pyproject.toml
[project]
name = "django-server"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.9.15,<4.0"
dependencies = [
    "django (>=3.2.13)",
    "python-decouple (>=3.8)",
    "psycopg2-binary (>=2.9.9)",
    "djangorestframework (>=3.15.1)",
    "pydantic (>=2.5.3)",
    "django-rest-knox (>=5.0.2)",
    "smsaero-api (>=3.0.0)",
    "structlog (>=24.4.0)",
    "ruff (>=0.11.2,<0.12.0)",
    "mypy (>=1.15.0,<2.0.0)",
    "django-stubs (>=5.1.3,<6.0.0)",
    "djangorestframework-types (==0.8.0)",
    "black (>=25.1.0,<26.0.0)",
    "django-migration-linter (>=5.1.0,<6.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disable_error_code = ["var-annotated", "import-untyped"]

[tool.black]
line-length = 120
target-version = ['py39']
skip-string-normalization = true
color = true
exclude = '''
/(
  | migrations
)/
'''
